+++ src/systems/ec/implementations/react/components/accordion2/src/Accordion2.jsx
