+++ src/systems/ec/implementations/vanilla/packages/ec-component-accordion/index.js
@@ -1,7 +1,6 @@
 import { Accordion } from './ec-component-accordion';
 
 export { Accordion } from './ec-component-accordion';
-export { Accordion2 } from './ec-component-accordion2';
 
 // Backward compatible: export Accordion by default
 export default Accordion;
