+++ src/systems/ec/implementations/vanilla/packages/ec-component-accordion2/README.md
@@ -1,0 +1,1 @@
+# Accordion 2
