+++ src/systems/ec/implementations/vanilla/packages/ec-component-accordion2/ec-component-accordion2.js
