+++ src/systems/ec/implementations/vanilla/packages/ec-component-accordion2/index.js
@@ -1,0 +1,5 @@
+import { Accordion2 } from './ec-component-accordion2';
+
+export { Accordion2 } from './ec-component-accordion2';
+
+export default Accordion2;
