+++ src/systems/ec/implementations/vanilla/packages/ec-component-accordion2/package.json
@@ -1,16 +1,17 @@
 {
-  "name": "@ecl/ec-component-accordion",
+  "name": "@ecl/ec-component-accordion2",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "2.8.0",
-  "description": "ECL EC Accordion",
+  "description": "ECL EC Accordion 2",
   "main": "index.js",
   "module": "index.js",
-  "style": "ec-component-accordion.scss",
-  "sass": "ec-component-accordion.scss",
+  "style": "ec-component-accordion2.scss",
+  "sass": "ec-component-accordion2.scss",
   "dependencies": {
     "@ecl/ec-base": "^2.8.0",
-    "@ecl/ec-component-button": "^2.8.0"
+    "@ecl/ec-component-button": "^2.8.0",
+    "@ecl/ec-component-icon": "^2.8.0"
   },
   "publishConfig": {
     "access": "public"
