+++ src/systems/ec/implementations/vanilla/packages/ec-preset-dev/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@ecl/ec-base": "^2.8.0",
     "@ecl/ec-component-accordion": "^2.8.0",
+    "@ecl/ec-component-accordion2": "^2.8.0",
     "@ecl/ec-component-blockquote": "^2.8.0",
     "@ecl/ec-component-breadcrumb": "^2.8.0",
     "@ecl/ec-component-button": "^2.8.0",
