+++ src/systems/ec/implementations/vanilla/packages/ec-preset-dev/src/ec-preset-dev-print.scss
@@ -41,7 +41,7 @@
 
 // Organisms
 @import '@ecl/ec-component-accordion/ec-component-accordion-print';
-@import '@ecl/ec-component-accordion/ec-component-accordion2-print';
+@import '@ecl/ec-component-accordion2/ec-component-accordion2-print';
 @import '@ecl/ec-component-gallery/ec-component-gallery-print';
 @import '@ecl/ec-component-page-header/ec-component-page-header-print';
 @import '@ecl/ec-component-site-header/ec-component-site-header-print';
