+++ src/systems/ec/implementations/vanilla/packages/ec-preset-dev/src/ec-preset-dev.js
@@ -1,6 +1,7 @@
 // Export components
 
 export * from '@ecl/ec-component-accordion';
+export * from '@ecl/ec-component-accordion2';
 export * from '@ecl/ec-component-breadcrumb';
 // export * from '@ecl/ec-component-carousel';
 export * from '@ecl/ec-component-contextual-navigation';
