+++ src/systems/ec/implementations/vanilla/packages/ec-preset-dev/src/ec-preset-dev.scss
@@ -77,7 +77,7 @@
 
 // Organisms
 @import '@ecl/ec-component-accordion/ec-component-accordion';
-@import '@ecl/ec-component-accordion/ec-component-accordion2';
+@import '@ecl/ec-component-accordion2/ec-component-accordion2';
 @import '@ecl/ec-component-gallery/ec-component-gallery';
 // @import '@ecl/ec-component-listing/ec-component-listing';
 // @import '@ecl/ec-component-navigation-list/ec-component-navigation-list';
