+++ src/systems/eu/implementations/react/components/accordion/index.js
@@ -1,5 +1,3 @@
 export default from './src/Accordion';
 export Accordion from './src/Accordion';
 export AccordionItem from './src/AccordionItem';
-export Accordion2 from './src/Accordion2';
-export Accordion2Item from './src/Accordion2Item';
