+++ src/systems/eu/implementations/react/components/accordion/package.json
@@ -8,8 +8,7 @@
   "main": "index.js",
   "module": "index.js",
   "dependencies": {
-    "@ecl/eu-react-component-button": "^2.8.0",
-    "@ecl/eu-react-component-icon": "^2.8.0"
+    "@ecl/eu-react-component-button": "^2.8.0"
   },
   "devDependencies": {
     "@ecl/eu-component-accordion": "^2.8.0",
