+++ src/systems/eu/implementations/react/components/accordion2/index.js
@@ -1,0 +1,3 @@
+export default from './src/Accordion2';
+export Accordion2 from './src/Accordion2';
+export Accordion2Item from './src/Accordion2Item';
