+++ src/systems/eu/implementations/react/components/accordion2/package.json
@@ -1,18 +1,18 @@
 {
   "private": true,
-  "name": "@ecl/ec-react-component-accordion",
+  "name": "@ecl/eu-react-component-accordion2",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "2.8.0",
-  "description": "ECL EC React Accordion",
+  "description": "ECL EU React Accordion 2",
   "main": "index.js",
   "module": "index.js",
   "dependencies": {
-    "@ecl/ec-react-component-button": "^2.8.0"
+    "@ecl/eu-react-component-icon": "^2.8.0"
   },
   "devDependencies": {
-    "@ecl/ec-component-accordion": "^2.8.0",
-    "@ecl/ec-specs-accordion": "^2.8.0"
+    "@ecl/eu-component-accordion2": "^2.8.0",
+    "@ecl/eu-specs-accordion2": "^2.8.0"
   },
   "peerDependencies": {
     "classnames": "^2.2.6",
