+++ src/systems/eu/implementations/react/components/accordion2/stories/Index.jsx
@@ -4,9 +4,9 @@ import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 import StoryWrapper from '@ecl/story-wrapper';
 
-import demoContent from '@ecl/eu-specs-accordion/demo/data';
+import demoContent from '@ecl/eu-specs-accordion2/demo/data';
 
-import { Accordion2 as VanillaAccordion } from '@ecl/eu-component-accordion';
+import { Accordion2 as VanillaAccordion } from '@ecl/eu-component-accordion2';
 
 import Accordion from '../src/Accordion2';
 import AccordionItem from '../src/Accordion2Item';
