+++ src/systems/eu/implementations/vanilla/packages/eu-component-accordion/index.js
@@ -1,7 +1,6 @@
 import { Accordion } from './eu-component-accordion';
 
 export { Accordion } from './eu-component-accordion';
-export { Accordion2 } from './eu-component-accordion2';
 
 // Backward compatible: export Accordion by default
 export default Accordion;
