+++ src/systems/eu/implementations/vanilla/packages/eu-component-accordion/package.json
@@ -10,8 +10,7 @@
   "sass": "eu-component-accordion.scss",
   "dependencies": {
     "@ecl/eu-base": "^2.8.0",
-    "@ecl/eu-component-button": "^2.8.0",
-    "@ecl/eu-component-icon": "^2.8.0"
+    "@ecl/eu-component-button": "^2.8.0"
   },
   "publishConfig": {
     "access": "public"
