+++ src/systems/eu/implementations/vanilla/packages/eu-component-accordion2/README.md
@@ -1,0 +1,1 @@
+# Accordion 2
