+++ src/systems/eu/implementations/vanilla/packages/eu-component-accordion2/eu-component-accordion2.scss
