+++ src/systems/eu/implementations/vanilla/packages/eu-component-accordion2/index.js
@@ -1,0 +1,5 @@
+import { Accordion2 } from './eu-component-accordion2';
+
+export { Accordion2 } from './eu-component-accordion2';
+
+export default Accordion2;
