+++ src/systems/eu/implementations/vanilla/packages/eu-component-accordion2/package.json
@@ -1,16 +1,17 @@
 {
-  "name": "@ecl/ec-component-accordion",
+  "name": "@ecl/eu-component-accordion2",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "2.8.0",
-  "description": "ECL EC Accordion",
+  "description": "ECL EU Accordion 2",
   "main": "index.js",
   "module": "index.js",
-  "style": "ec-component-accordion.scss",
-  "sass": "ec-component-accordion.scss",
+  "style": "eu-component-accordion2.scss",
+  "sass": "eu-component-accordion2.scss",
   "dependencies": {
-    "@ecl/ec-base": "^2.8.0",
-    "@ecl/ec-component-button": "^2.8.0"
+    "@ecl/eu-base": "^2.8.0",
+    "@ecl/eu-component-button": "^2.8.0",
+    "@ecl/eu-component-icon": "^2.8.0"
   },
   "publishConfig": {
     "access": "public"
