+++ src/systems/eu/implementations/vanilla/packages/eu-preset-dev/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@ecl/eu-base": "^2.8.0",
     "@ecl/eu-component-accordion": "^2.8.0",
+    "@ecl/eu-component-accordion2": "^2.8.0",
     "@ecl/eu-component-blockquote": "^2.8.0",
     "@ecl/eu-component-breadcrumb": "^2.8.0",
     "@ecl/eu-component-button": "^2.8.0",
