+++ src/systems/eu/implementations/vanilla/packages/eu-preset-dev/src/eu-preset-dev-print.scss
@@ -40,7 +40,7 @@
 
 // Organisms
 @import '@ecl/eu-component-accordion/eu-component-accordion-print';
-@import '@ecl/eu-component-accordion/eu-component-accordion2-print';
+@import '@ecl/eu-component-accordion2/eu-component-accordion2-print';
 @import '@ecl/ec-component-gallery/ec-component-gallery-print';
 @import '@ecl/eu-component-page-header/eu-component-page-header-print';
 @import '@ecl/eu-component-site-header/eu-component-site-header-print';
