+++ src/systems/eu/implementations/vanilla/packages/eu-preset-dev/src/eu-preset-dev.js
@@ -1,6 +1,7 @@
 // Export components
 
 export * from '@ecl/eu-component-accordion';
+export * from '@ecl/eu-component-accordion2';
 export * from '@ecl/eu-component-breadcrumb';
 // export * from '@ecl/eu-component-carousel';
 export * from '@ecl/eu-component-contextual-navigation';
