+++ src/systems/eu/implementations/vanilla/packages/eu-preset-dev/src/eu-preset-dev.scss
@@ -65,7 +65,7 @@
 
 // Organisms
 @import '@ecl/eu-component-accordion/eu-component-accordion';
-@import '@ecl/eu-component-accordion/eu-component-accordion2';
+@import '@ecl/eu-component-accordion2/eu-component-accordion2';
 @import '@ecl/eu-component-gallery/eu-component-gallery';
 // @import '@ecl/eu-component-listing/eu-component-listing';
 // @import '@ecl/eu-component-navigation-list/eu-component-navigation-list';
