+++ src/systems/eu/specs/components/accordion2/demo/data.js
@@ -1,0 +1,38 @@
+// Simple content for demo
+module.exports = {
+  items: [
+    {
+      id: 'accordion-example',
+      level: 3,
+      toggle: {
+        label:
+          'Delivery of last pending proposals, a common Destiny of unity, the hour of European Democracy',
+        iconShape: 'general--digital',
+      },
+      content:
+        'The College of Commissioners held today the first weekly meeting of 2019 which was devoted to discussing the challenges of this new year. Commissioners used the opportunity to take stock and discuss the year ahead, including the European elections in May and other important milestones ahead.',
+    },
+    {
+      id: 'accordion-example2',
+      level: 3,
+      toggle: {
+        label:
+          'Spring 2019 Economic Forecast: Growth continues at a more moderate pace',
+        iconShape: 'general--growth',
+      },
+      content:
+        'The European economy is forecast to continue expanding for the seventh year in a row in 2019, with real GDP expected to grow in all EU Member States. As global uncertainties continue to weigh, domestic dynamics are set to support the European economy. Growth is expected to gather pace again next year.',
+    },
+    {
+      id: 'accordion-example3',
+      level: 3,
+      toggle: {
+        label:
+          'Delivery of last pending proposals, a common Destiny of unity, the hour of European Democracy',
+        iconShape: 'general--regulation',
+      },
+      content:
+        'In the modern global economy trade is essential for growth, jobs and competiveness, and the EU is committed to maintaining an open and rules-based trading system. With the rising threat of protectionism and weakened commitment of large players to global trade governance, the EU must take the lead.',
+    },
+  ],
+};
