+++ src/systems/eu/specs/components/accordion2/package.json
@@ -1,17 +1,9 @@
 {
-  "name": "@ecl/ec-component-accordion",
+  "name": "@ecl/eu-specs-accordion2",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "2.8.0",
-  "description": "ECL EC Accordion",
-  "main": "index.js",
-  "module": "index.js",
-  "style": "ec-component-accordion.scss",
-  "sass": "ec-component-accordion.scss",
-  "dependencies": {
-    "@ecl/ec-base": "^2.8.0",
-    "@ecl/ec-component-button": "^2.8.0"
-  },
+  "description": "ECL EU Accordion 2 Specs",
   "publishConfig": {
     "access": "public"
   },
