+++ src/website/src/pages/ec/components/accordion/docs/code.mdx
@@ -4,9 +4,9 @@ order: 2
 ---
 
 import Playground from '@ecl/website-components/Playground';
-import { Accordion2, Accordion2Item } from '@ecl/ec-react-component-accordion';
+import { Accordion2, Accordion2Item } from '@ecl/ec-react-component-accordion2';
 
-import demoContent from '@ecl/ec-specs-accordion/demo/data';
+import demoContent from '@ecl/ec-specs-accordion2/demo/data';
 
 <Playground
   system="ec"
