+++ src/website/src/pages/eu/components/accordion/docs/code.mdx
@@ -4,9 +4,9 @@ order: 2
 ---
 
 import Playground from '@ecl/website-components/Playground';
-import { Accordion2, Accordion2Item } from '@ecl/eu-react-component-accordion';
+import { Accordion2, Accordion2Item } from '@ecl/eu-react-component-accordion2';
 
-import demoContent from '@ecl/eu-specs-accordion/demo/data';
+import demoContent from '@ecl/eu-specs-accordion2/demo/data';
 
 <Playground
   system="eu"
